compose_textarea: Move `position_inside_code_block` from compose_ui.
This is a pure move of the existing `position_inside_code_block`
helper from `compose_ui.ts` to `compose_textarea.ts`, with no change
to its implementation, to avoid a circular dependency in a later
commit that uses it from the message-splitting code. `compose_ui`
now delegates to the moved function.

Fixes part of #28626.
